Refactor auth and audio helpers into modules
